:recycle: refactor: cria NavbarEmpresa e unifica navbars de empresa #162
- Cria componente NavbarEmpresa parametrizável (showSearch, initialSearchValue)
- Refatora EmpresaDashboard para usar NavbarEmpresa
- Refatora BuscaCandidatos para usar NavbarEmpresa
- Organiza NavbarEmpresa dentro de components/Navbar/
- Remove ~90 linhas de código duplicado
- Reduz bundle JS em ~1.4 kB
