+++ frontend/src/pages/BuscaCandidatos/BuscaCandidatos.module.css
@@ -3,55 +3,10 @@
   background-color: #f5f5f5;
 }
 
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #2c3e50;
-  color: #fff;
-  padding: 10px 20px;
-  position: sticky;
-  top: 0;
-  z-index: 10;
-}
-
-.navbar h1 {
-  font-size: 24px;
-  margin: 0;
-}
-
-.navbarMenu {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  gap: 20px;
-}
-
-.logoutButton {
-  background: none;
-  border: none;
-  color: #fff;
-  cursor: pointer;
-  padding: 10px;
-  border-radius: 5px;
-  transition: background-color 0.3s;
-}
-
-.logoutButton:hover {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.logoutButton svg {
-  width: 20px;
-  height: 20px;
-}
-
-/* Search Container */
-.searchContainer {
+/* Header Bar with Back Button */
+.headerBar {
   display: flex;
   align-items: center;
-  justify-content: space-between;
   padding: 15px 20px;
   background-color: white;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
@@ -71,32 +26,6 @@
 
 .backButton:hover {
   background-color: #f0f0f0;
-}
-
-.searchForm {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  flex: 1;
-  max-width: 600px;
-  margin: 0 20px;
-}
-
-.searchIcon {
-  font-size: 1.2rem;
-}
-
-.searchInput {
-  flex: 1;
-  padding: 10px 15px;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-}
-
-.searchInput:focus {
-  outline: none;
-  border-color: #ffe100;
 }
 
 .searchButton {
